refactor: moves component to correct folder
