+++ src/components/atomic-swap/BalanceComponent.js
